20260209_LangExtract/tests/inference_test/pyproject.toml
[project]
name = "inference-test"
version = "0.1.0"
description = "Test scripts for LLM inference endpoints"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

